+++ Source/Device.generated.swift
@@ -236,6 +236,8 @@ public enum Device {
     case iPhone18Pro
     /// Device is an [iPhone 18 Pro Max](https://support.apple.com/en-us/148591)
     case iPhone18ProMax
+    /// Provisional iPhone Duo; identifier, display metrics, and other capabilities await verification.
+    case iPhoneDuo
     /// Device is an [iPad 2](https://support.apple.com/kb/SP622)
     ///
     /// ![Image](https://support.apple.com/library/APPLE/APPLECARE_ALLGEOS/SP622/SP622_01-ipad2-mul.png)
@@ -638,6 +640,7 @@ public enum Device {
       case "iPhone18,5": return iPhone17e
       case "iPhone19,2": return iPhone18Pro
       case "iPhone19,3", "iPhone19,7": return iPhone18ProMax
+      case "placeholder:iPhoneDuo": return iPhoneDuo
       case "iPad2,1", "iPad2,2", "iPad2,3", "iPad2,4": return iPad2
       case "iPad3,1", "iPad3,2", "iPad3,3": return iPad3
       case "iPad3,4", "iPad3,5", "iPad3,6": return iPad4
@@ -744,6 +747,19 @@ public enum Device {
     #endif
   }
 
+  /// Returns whether the device has a foldable display, including simulators.
+  public var isFoldable: Bool {
+    #if os(iOS)
+    switch self {
+    case .iPhoneDuo: return true
+    case .simulator(let model): return model.isFoldable
+    default: return false
+    }
+    #else
+    return false
+    #endif
+  }
+
   /// Get the real device from a device.
   /// If the device is a an iPhone8Plus simulator this function returns .iPhone8Plus (the real device).
   /// If the parameter is a real device, this function returns just that passed parameter.
@@ -818,6 +834,7 @@ public enum Device {
         case .iPhone17e: return 6.1
         case .iPhone18Pro: return 6.3
         case .iPhone18ProMax: return 6.9
+        case .iPhoneDuo: return -1
         case .iPad2: return 9.7
         case .iPad3: return 9.7
         case .iPad4: return 9.7
@@ -969,6 +986,7 @@ public enum Device {
       case .iPhone17e: return (width: 9, height: 19.5)
       case .iPhone18Pro: return (width: 9, height: 19.5)
       case .iPhone18ProMax: return (width: 9, height: 19.5)
+      case .iPhoneDuo: return (width: -1, height: -1)
       case .iPad2: return (width: 3, height: 4)
       case .iPad3: return (width: 3, height: 4)
       case .iPad4: return (width: 3, height: 4)
@@ -1073,7 +1091,7 @@ public enum Device {
 
     /// All iPhones
     public static var allPhones: [Device] {
-      return [.iPhone4, .iPhone4s, .iPhone5, .iPhone5c, .iPhone5s, .iPhone6, .iPhone6Plus, .iPhone6s, .iPhone6sPlus, .iPhone7, .iPhone7Plus, .iPhoneSE, .iPhone8, .iPhone8Plus, .iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPhoneSE2, .iPhone12, .iPhone12Mini, .iPhone12Pro, .iPhone12ProMax, .iPhone13, .iPhone13Mini, .iPhone13Pro, .iPhone13ProMax, .iPhoneSE3, .iPhone14, .iPhone14Plus, .iPhone14Pro, .iPhone14ProMax, .iPhone15, .iPhone15Plus, .iPhone15Pro, .iPhone15ProMax, .iPhone16, .iPhone16Plus, .iPhone16Pro, .iPhone16ProMax, .iPhone16e, .iPhone17, .iPhone17Pro, .iPhone17ProMax, .iPhoneAir, .iPhone17e, .iPhone18Pro, .iPhone18ProMax]
+      return [.iPhone4, .iPhone4s, .iPhone5, .iPhone5c, .iPhone5s, .iPhone6, .iPhone6Plus, .iPhone6s, .iPhone6sPlus, .iPhone7, .iPhone7Plus, .iPhoneSE, .iPhone8, .iPhone8Plus, .iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPhoneSE2, .iPhone12, .iPhone12Mini, .iPhone12Pro, .iPhone12ProMax, .iPhone13, .iPhone13Mini, .iPhone13Pro, .iPhone13ProMax, .iPhoneSE3, .iPhone14, .iPhone14Plus, .iPhone14Pro, .iPhone14ProMax, .iPhone15, .iPhone15Plus, .iPhone15Pro, .iPhone15ProMax, .iPhone16, .iPhone16Plus, .iPhone16Pro, .iPhone16ProMax, .iPhone16e, .iPhone17, .iPhone17Pro, .iPhone17ProMax, .iPhoneAir, .iPhone17e, .iPhone18Pro, .iPhone18ProMax, .iPhoneDuo]
     }
 
     /// All iPads
@@ -1551,6 +1569,7 @@ public enum Device {
       case .iPhone17e: return 460
       case .iPhone18Pro: return 460
       case .iPhone18ProMax: return 460
+      case .iPhoneDuo: return nil
       case .iPad2: return 132
       case .iPad3: return 264
       case .iPad4: return 264
@@ -1734,6 +1753,7 @@ extension Device: CustomStringConvertible {
       case .iPhone17e: return "iPhone 17e"
       case .iPhone18Pro: return "iPhone 18 Pro"
       case .iPhone18ProMax: return "iPhone 18 Pro Max"
+      case .iPhoneDuo: return "iPhone Duo"
       case .iPad2: return "iPad 2"
       case .iPad3: return "iPad (3rd generation)"
       case .iPad4: return "iPad (4th generation)"
@@ -1904,6 +1924,7 @@ extension Device: CustomStringConvertible {
       case .iPhone17e: return "iPhone 17e"
       case .iPhone18Pro: return "iPhone 18 Pro"
       case .iPhone18ProMax: return "iPhone 18 Pro Max"
+      case .iPhoneDuo: return "iPhone Duo"
       case .iPad2: return "iPad 2"
       case .iPad3: return "iPad (3rd generation)"
       case .iPad4: return "iPad (4th generation)"
@@ -2634,6 +2655,7 @@ extension Device {
       case .iPhone17e: return .a19
       case .iPhone18Pro: return .a20Pro
       case .iPhone18ProMax: return .a20Pro
+      case .iPhoneDuo: return .a20Pro
       case .iPad2: return .a5
       case .iPad3: return .a5X
       case .iPad4: return .a6X